requirements.txt
PyQt6
google-genai
cryptography
prompt-toolkit
rich
tomli
websocket-client
